+++ services/mongo.js
@@ -1,6 +1,6 @@
-module.exports = function(config, app) {
-  const { MongoClient } = require('mongodb')
-  let dbName = `${app.name}_${app.env}`
-  let defaultConnectionUrl = `mongodb://localhost:27017/${dbName}`
-  return MongoClient.connect(config.get('db.endpoint', defaultConnectionUrl))
+module.exports = function(config, app, mongoConnectionFactory) {
+  const dbName = `${app.name}_${app.env}`
+  const defaultConnectionUrl = `mongodb://localhost:27017/${dbName}`
+  const connectionString = config.get('db.endpoint', defaultConnectionUrl)
+  return mongoConnectionFactory.connectToDB(connectionString)
 }